Deal with orphaned archive buffers
If the buffer associated with an archive is killed (for instance
because it was opened with emacsclient and its frame was closed), but
some buffers associated with the constituent files remain live, these
buffers are "orphans".  Unless they're dealt with specially (as patched
here) they'll result in the filter not working.

Now, all the orphans will be grouped together (irrespective of archive type).